+++ application/controllers/photos/Delete.php
@@ -4,7 +4,8 @@
 */
 class Delete extends CI_Controller
 {
-    private $media_path;
+    private $media_path,
+            $permissions = [''];
 
     function __construct()
     {
@@ -28,7 +29,14 @@ class Delete extends CI_Controller
                 "status" => "error",
                 "message" => "Internal error has occured."
             ];
-
+            if(!in_array('all',$this->permissions) && !in_array('photo_delete',$this->permissions))
+            {
+                $response['code'] = '403';
+                $response['message'] = "You're not authorized to perform this action.";
+                header("Content-Type: application/json");
+                echo json_encode($response);
+                exit();
+            }
             if($deleted = $this->m_photo->delete($id))
             {
                 $response["status"] = "ok";
